requirements.txt
alembic==1.4.3
aniso8601==8.1.0
appdirs==1.4.3
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
html5lib==1.0.1
httpie==2.3.0
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
migrate==0.3.8
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
Pygments==2.7.3
pyparsing==2.4.6
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.5
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.22
urllib3==1.25.8
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
yml==0.0.1
